Cgroups2: Reduce allocations
There's a couple spots we can get some free wins when reading single
uint values from files. We were using `os.ReadFile` in one spot which
needs to allocate to return the byte slice containing the data for the
file. We can open the file and read in N bytes instead so it stays on
the stack.

Similarly we were using `io.ReadAll` in another instance which has to
allocate to return the byte slice read from the reader.

Signed-off-by: Danny Canter <danny@dcantah.dev>