Cgroups2: Reduce allocations for manager.Stat
This change works towards bringing down allocations for manager.Stat().
We can shave off 3 by providing a size upfront for the map we store stats
in, as otherwise it'll resize at runtime as we're storing the 40+ stats
inside. This also reworks the type of the map, as storing strings or uints
in an interface{} should allocate as well. This shaves off 9 allocs, and
I'm positive there's more that can be done here, but this is a decent start.

Signed-off-by: Danny Canter <danny@dcantah.dev>